Fix: warnings

+++ VTPG_Common.m
@@ -27,11 +27,11 @@ static NSString* VTPGStringFromBoolOrCharValue(BOOL boolOrCharvalue) {
 
 static NSString *VTPGStringFromFourCharCodeOrUnsignedInt32(FourCharCode fourcc) {
 	return [NSString stringWithFormat:@"%u ('%c%c%c%c')",
-										fourcc,
-										(fourcc >> 24) & 0xFF,
-										(fourcc >> 16) & 0xFF,
-										(fourcc >> 8) & 0xFF,
-										fourcc & 0xFF];
+										(unsigned int)fourcc,
+										(unsigned int)((fourcc >> 24) & 0xFF),
+										(unsigned int)((fourcc >> 16) & 0xFF),
+										(unsigned int)((fourcc >> 8) & 0xFF),
+										(unsigned int)(fourcc & 0xFF)];
 }
 
 static NSString *StringFromNSDecimalWithCurrentLocal(NSDecimal dcm) {
@@ -76,7 +76,7 @@ NSString * VTPG_DDToStringFromTypeAndValue(const char * typeCode, void * value) 
 	IF_TYPE_MATCHES_INTERPRET_WITH_FORMAT(unsigned short,@"%hu");
 	IF_TYPE_MATCHES_INTERPRET_WITH_FORMAT(int,@"%i");
 	IF_TYPE_MATCHES_INTERPRET_WITH_FORMAT(unsigned, @"%u");
-	IF_TYPE_MATCHES_INTERPRET_WITH_FORMAT(long,@"%i");
+	IF_TYPE_MATCHES_INTERPRET_WITH_FORMAT(long,@"%ld");
 	IF_TYPE_MATCHES_INTERPRET_WITH_FORMAT(long double,@"%Lf"); //WARNING on older versions of OS X, @encode(long double) == @encode(double)
 	
 	//C-strings
